Fix: Listings that show up have markers now
I had to refactor part of the create markers in order for it to be
reused once the zoom button was pressed so that each location that is
shown on the list would also have a corresponding marker that would also
show up. Earlier the list did not have a marker that was created.